Create spiderBlastParser.py
This script utilizes the BioPython built-in Blast XML parser to extract information about top hits in a blast query and write them to a tab-separated text file.    Although the parser makes this relatively straightforward, it is far from perfect.  In order to get readable results, I needed to do some additional string parsing inside some of the xml tags.